mcp-server/requirements.txt
# MCP Framework
mcp>=1.0.0

# HTTP/SSE Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0

# Couchbase SDK
couchbase>=4.2.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.1.0

# Authentication
pyjwt>=2.8.0
python-multipart>=0.0.6

# HTTP Client for Ollama
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
loguru>=0.7.2

# Vector operations
numpy>=1.24.0
